Update table.py and related docs
